chore : erase unused variable isSpeech
